move memo structs into composable traits
